+++ src/front/js/pages/Groups.js
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { Container, Row, Col, Button, Modal, Form, Dropdown } from 'react-bootstrap';
 import { Context } from "../store/appContext"
 import { Link , useNavigate} from "react-router-dom";
+import GroupInfo from "../component/GroupInfo";
 
 // notes to self:
 //     X: need add group button that leads to create group Modal
@@ -22,9 +23,6 @@ const Groups = (props) => {
 
     const handleAddGroup = ()=> {
         actions.showGroupModal(true);
-    }
-    const handleEditMember = () => {
-        actions.showEditMemberModal(true)
     }
     const handleCloseModal = ()=> {
         actions.hideGroupModal(false)
@@ -49,40 +47,9 @@ const Groups = (props) => {
                         <button className='expense-btn' onClick={handleAddGroup}>Create New Group</button>
                     </Col>
                 </Row>
-                <div className='border border-2 border-dark rounded p-3'>
-                    <Row>
-                        <Col>
-                            <h3>{store.userGroups[0].name}</h3>
-                        </Col>
-                        <Col className='d-flex flex-row-reverse align-items-center'>
-                            <button className='expense-btn' onClick={handleEditMember}>Edit Group</button>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <h6>Group Member 1, Group Member 2, Group Member 3</h6>
-                    </Row>
-                    <Row>
-                        <Col>
-                            <h5>Shared Expenses</h5>
-                        </Col>
-                        <Col className='d-flex flex-row-reverse align-items-center'>
-                            <button className='expense-btn'>Add Expense</button>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <h6 className="ms-2">Expense 1</h6>
-                        <h6 className="ms-2">Expense 2</h6>
-                        <h6 className="ms-2">Expense 3</h6>
-                    </Row>
-                    <Row>
-                        <Col>
-                            <h5>Balance to User</h5>
-                        </Col>
-                        <Col>
-                            <h5 className='d-flex flex-row-reverse'>$0.00</h5>
-                        </Col>
-                    </Row>                    
-                </div>
+                {store.userGroups.map((group) =>{
+                    return <GroupInfo group={group}/>
+                })}
             </Container>
             }
